requirements.txt
altair==6.0.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.61.1
gitdb==4.0.12
gitpython==3.1.46
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
kagglehub==1.0.0
kagglesdk==0.1.15
kiwisolver==1.4.9
loguru==0.7.3
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.16.0
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
plotly==6.5.2
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
prompt-toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
ruff==0.15.2
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.54.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
